LayoutTests/imported/w3c/web-platform-tests/css/css-scoping/has-slotted-changing-002.tentative-expected.txt


FAIL empty node is blue assert_equals: expected "rgb(0, 0, 255)" but got "rgb(0, 255, 0)"
FAIL manually assigning a text node is green, emptying assignment is blue assert_equals: expected "rgb(0, 0, 255)" but got "rgb(0, 255, 0)"
FAIL manually assigning a Element is green, emptying assignment is blue assert_equals: expected "rgb(0, 0, 255)" but got "rgb(0, 255, 0)"
